fix: require resolve (#4356)
* fix: require resolve

* fix: error

* fix: skip realpath

* fix: fs.realpath

+++ packages/core/utils/plugin-symlink.js
@@ -1,5 +1,5 @@
 const { dirname, resolve } = require('path');
-const { readFile, writeFile, readdir, symlink, unlink, mkdir, stat } = require('fs').promises;
+const { realpath, readFile, writeFile, readdir, symlink, unlink, mkdir, stat } = require('fs').promises;
 
 async function getStoragePluginNames(target) {
   const plugins = [];
@@ -76,6 +76,10 @@ async function createDevPluginSymLink(pluginName) {
     }
     const link = resolve(nodeModulesPath, pluginName);
     if (await fsExists(link)) {
+      const real = await realpath(link);
+      if (real === resolve(packagePluginsPath, pluginName)) {
+        return;
+      }
       await unlink(link);
     }
     await symlink(resolve(packagePluginsPath, pluginName), link, 'dir');
